refactor: 将id和classId类型从Integer改为Long以适配更大范围的数值

+++ src/main/java/com/furinafans/StuSys/controller/CourseController.java
@@ -27,7 +27,7 @@ public class CourseController {
 
     //插入课程
     @PostMapping
-    public Result<Integer> addCourse(@RequestBody @Valid CourseDTO courseDTO) {
+    public Result<Long> addCourse(@RequestBody @Valid CourseDTO courseDTO) {
         Course course = Course.builder()
                 .name(courseDTO.getName().strip())
                 .code(courseDTO.getCode().strip())
